Indicate if the model is placeable or not
When the player tries to place the demo on an occupied ground, the ground will be read indicating it's not possible to place the model on it. This feature is still in progress since it only take into consideration the appliedX and appliedY (the block the user is hovering over) and ignore the reset of the blocks that the building should occupy.