Interact and collect buildings
The left hand is now only dedicated for interacting with buildings. When the left hand raycast interact with the buildings, their color changes to redish one. When the Activate button is clicked while hovering over a gold mine, the gold inside of gold mine will be collected and saved int he gold bank which is the arc around the left hand