Demo adder edited, bug fixed.
The UI will be destroyed after adding the last building which the UI represents in stock. Sounds to indicate the clicks and demo added confirmation has been added.

Bug fixed:
The BattleProgress script has the spawningStarted variable as static which will not change its state even when the scene is reloaded. This caused the scene to be reloaded again and again. This has been fixed by creating a reset state method which resets the variable to false.