UIs fixed.
Gold bank has now a gold update call back register system, which allows the Upgrade menu to update whether an item is buyable or not in real time. The same applies for the barracks warriors price.

The Barracks production UI has the ability to cancel the production of a UI now.